Moving globals to Test Runner class
