+++ packages/app/src/sandbox/eval/transpiled-module.js
@@ -18,7 +18,7 @@ import type { WarningStructure } from './transpilers/utils/worker-warning-handle
 import resolveDependency from './loaders/dependency-resolver';
 import evaluate from './loaders/eval';
 
-import * as testGlobals from './tests/jest-lite';
+import TestRunner from './tests/jest-lite';
 
 import Manager from './manager';
 import HMR from './hmr';
@@ -664,7 +664,7 @@ export default class TranspiledModule {
         require,
         this.compilation,
         manager.envVariables,
-        testGlobals
+        TestRunner.testGlobals()
       );
 
       const hmrConfig = this.hmrConfig;
